Win warnings (#3625)
* Fixes for GCC 11

* Remove OCI.

* Update python filter doc.

* Deploy for 2.3-maintenance.

* Doc fix.

* Doc update.

* Update conda-forge pdal version to 2.3 on download docpage (#3504)

* Check that a file is mapped before attempting to unmap.

* Eliminate a bunch of Windows warnings.

* Remove STDCALL definition since it's only used one place.

* Add missed header.

* Eliminate STDCALL.

+++ filters/private/expr/Lexer.cpp
@@ -203,7 +203,7 @@ Token Lexer::number()
     char *end;
     double v = strtod(start, &end);
     m_pos = end - m_buf.data();
-    int len = end - start;
+    int len = (int)(end - start);
     return Token(TokenType::Number, m_tokPos, m_pos, m_buf.substr(m_tokPos, len), v);
 }
